Roll in baseline changes from GDB 6.7 and binutils 2.23
This is too major a step - from a code base in part 7 years old - to do a
clean git cherry-pick. Instead we have brought the baseline in from the
binutils-2.23 branch for AVR (for bfd, opcodes and include) and from the GDB
6.7 branch (for gdb). There is in any case no true log for those branches,
since they exist only as patch sets.

Consult the various ChangeLog.AVR32 files to glean more information regarding
the history of changes.

Note that this commit is not expected to build. It is a starting point from
which a buildable system will be constructed.